Move dynamodb table creation to slurm_builder
A dynamodb table is needed only if the scheduler is Slurm. The wrongly creation of dynamodb table for AWS Batch was added during the refactor of the pcluster3 code.
