+++ cli/tests/pcluster/templates/test_cluster_stack/test_head_node_dna_json/awsbatch-imds-secured-false.head-node.dna.json
@@ -10,7 +10,7 @@
     "cw_logging_enabled": "true",
     "dcv_enabled": "false",
     "dcv_port": "NONE",
-    "ddb_table": "{'Ref': 'DynamoDBTable'}",
+    "ddb_table": "NONE",
     "disable_hyperthreading_manually": "false",
     "dns_domain": "",
     "ebs_shared_dirs": "NONE,NONE,NONE,NONE,NONE",
